\name{nutriNetwork-package}
\alias{nutriNetwork-package}
\docType{package}
\title{
  Undirected Network for nutrition multivariate data
}

\description{
Statistical tool for learning the structure of direct associations among variables for 
			 continuous data, discrete data and mixed discrete-continuous data. The package is based
			 on the copula graphical model in Behrouzi and Wit (2017) <doi:10.1111/rssc.12287>. 
}

\author{Pariya Behrouzi \cr
Maintainers: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
1. Behrouzi, P., and Wit, E. C. (2019). Detecting epistatic selection with partially observed genotype data by using copula graphical models. Journal of the Royal Statistical Society: Series C (Applied Statistics), 68(1), 141-160. \cr
3. Behrouzi, P., and Wit, E. C. (2017c). netgwas: An R Package for Network-Based Genome-Wide Association Studies. arXiv preprint, arXiv:1710.01236.
}

\keyword{package}