\name{vfit}
\alias{vfit}
\docType{data}
\title{Baseline data from VFIT study}
\description{ 
A dietary study that includs dietary intake, physical performance, and muscle strength-related variables for 207 Dutch elderly people.
}

\usage{data(vfit)}
\format{
  The format is a matrix containing 29 variables for 207 participants.
}  

\details{
Participants of the V-Fit trial were recruited via personal letters sent to senior residencies, home care organisations, general practitioners and local advertisements. Eligible participants were aged 70 y and older, used care services, did not regularly exercise, had a BMI of less than 25.
}

\source{
Paw, M. J. C. A., de Jong, N., Schouten, E. G., Hiddink, G. J., & Kok, F. J. (2001). Physical exercise and/or enriched foods for functional improvement in frail, independently living elderly: a randomized controlled trial. Archives of physical medicine and rehabilitation, 82(6), 811-817.
}

\examples{
data(vfit)
head(vfit, n=3)
}

\keyword{datasets}
